backend/requirements.txt
python-multipart
einops==0.8.2
timm==1.0.24
torch
torchvision
numpy
Pillow
matplotlib
scipy
h5py
imageio
scikit-learn
scikit-image
opencv-python-headless
tqdm
transformers
runpod